gardalte: Create /data and /cache partitions on sdcard
The internal flash memory is just too small to be useful for Ubuntu
Touch. This commit adds a script that runs on recovery boot and will try
to resize the existing sdcard partition so there's 4GiB of unpartitioned
space, then create a 3GiB /data partition and a 1GiB /cache partition on
the sdcard. It will only do the repartitioning work if there's a single
partition on the sdcard. /data is assumed to be /dev/mmcblk1p2 and
/cache is /dev/mmcblk1p3 (plus /block/ in recovery because of ueventd /
udev differences).

There's a timeout of 1800 seconds for the script to run. If it's not
successful after 1800 seconds, recovery will boot up and you probably
don't have the partitions required to install Ubuntu Touch. Instead you
can use adb to check /tmp/ubuntu-sd-partitioner.log.